7894/7894/7894/expense_tracker/requirements.txt
Django==4.2.0
gunicorn==21.2.0
whitenoise==6.4.0
python-decouple==3.8
Pillow==10.0.0
groq==1.0.0
